+++ purchases/src/main/kotlin/com/revenuecat/purchases/galaxy/conversions/ProductTypeConversions.kt
@@ -1,0 +1,19 @@
+package com.revenuecat.purchases.galaxy.conversions
+
+import com.revenuecat.purchases.ProductType
+import com.revenuecat.purchases.common.LogIntent
+import com.revenuecat.purchases.common.log
+import com.revenuecat.purchases.galaxy.GalaxyStrings
+
+internal fun String.createRevenueCatProductTypeFromSamsungIAPTypeString(): ProductType {
+    return when (this.lowercase()) {
+        "item" -> ProductType.INAPP
+        "subscription" -> ProductType.SUBS
+        else -> {
+            log(LogIntent.GALAXY_WARNING) {
+                GalaxyStrings.UNKNOWN_GALAXY_IAP_TYPE_STRING.format(this)
+            }
+            ProductType.UNKNOWN
+        }
+    }
+}
